ModBus 레지스터 타입을 short로 변경 및 Raw 지원
ModBus 레지스터 관련 타입을 int/ushort에서 short(16비트 signed)로 통일하였습니다.
인터페이스 및 구현체의 메서드 시그니처를 모두 short로 변경하였고, Device 클래스의 내부 자료구조도 이에 맞게 수정했습니다.
레지스터를 byte 배열로 직접 읽는 ReadRawHoldingRegisters, ReadRawInputRegisters 메서드를 추가하여 Raw 데이터 접근 기능을 확장했습니다.
응답 처리에 short 값을 지원하는 AddWriteResponse 및 SetShort 메서드를 추가하여 전체적으로 데이터 타입 일관성을 강화했습니다.

+++ ModBusQ/IModBusServer.cs
@@ -30,10 +30,10 @@ public interface IModBusServer : IDisposable
 	bool RemoveDevice(int devId);
 
 	void SetCoils(int devId, int address, params bool[] values);
-	void SetHoldingRegisters(int devId, int address, params int[] values);
+	void SetHoldingRegisters(int devId, int address, params short[] values);
 
 	bool GetCoil(int devId, int address);
 	bool GetDiscreteInput(int devId, int address);
-	int GetHoldingRegister(int devId, int address);
-	int GetInputRegister(int devId, int address);
+	short GetHoldingRegister(int devId, int address);
+	short GetInputRegister(int devId, int address);
 }
